Fix Active Visit tag styling (#196)
Fix the styling of the contents of the patient banner both when the Active Visit tag is present and when it's absent. This involves fixing the alignment of the patient name and Active Visit tag so that they align horizontally at the same level as the top of the patient avatar.