Add default value for errorCopy translation (#228)
Adds an optional default value for the errorCopy translation in the `ErrorState` component. This
is in keeping with the convention we've established in this repo where we provide the translation key as well as the optional translation default value. This default value is what is going to be displayed in case translations don't work - as is the situation now for reasons I'm not sure of.